refactor(capture): split the log-activity composer to stay under the file-length cap
`entityref.tsx` and `logactivity.tsx` are both grandfathered over the 500-LOC
ceiling, frozen by the ratchet at 670 and 616; the assignee work grew each past
its frozen size, which the fe-file-length gate refuses — a waived file may
shrink but never grow.

Move the two new concepts into their own files rather than bump the waiver:

- `assigneepicker.tsx` — `useAssignableUserOptions` (the roster-minus-agents
  eligibility rule) and the `TaskAssigneeField` control that offers it. This
  returns `entityref.tsx` to its frozen 670.
- `activitybody.ts` — the `ActivityDraft` shape and its wire projection
  (`activityRequestBody`, `occurredInstant`, `KINDS_WITH_A_PERSON`), the pure
  POST-body logic with no React, so a test can pin the body without a component.

`logactivity.tsx` keeps the React composer alone and drops to 528. The picker
now takes a plain `active` boolean instead of the draft's kind, so it depends on
neither the draft type nor the composer. Behaviour is unchanged; the full
composer, worklist and leads suites stay green.
