+++ frontend/src/screens/entityref.tsx
@@ -212,36 +212,6 @@ export function useRoster(kind: RosterKind, enabled: boolean) {
   });
 }
 
-/** A workspace user as a picker option. */
-export type UserOption = { value: string; label: string };
-
-/**
- * The workspace's people as assignment options — everyone the roster carries
- * LESS agent seats.
- *
- * An agent seat is an Agent Runner identity, not a person: the server refuses
- * one as an assignee or owner (the activities module's
- * `ensureAssigneeCanHoldWork`), so offering it is a control whose only outcome
- * is a refusal. Every picker that hands work to somebody asks that same
- * question of the same roster, so it is spelled once here rather than re-derived
- * beside each one — a second copy is how the two would drift until one offered a
- * seat the other refused. A caller that OFFERS this list still owes its reader
- * `useRosterPartial` beside it, for the same reason `useRoster` does.
- *
- * `enabled` defers the roster walk for a caller mounted before it needs the
- * list — a composer whose assignee picker only applies to a task has no reason
- * to walk `/users` while a note is being written.
- */
-export function useAssignableUserOptions(enabled = true): UserOption[] {
-  const roster = useRoster("user", enabled);
-  return (roster.data ?? [])
-    .filter((entry) => !("is_agent" in entry && entry.is_agent))
-    .map((entry) => ({
-      value: entry.id,
-      label: ("display_name" in entry ? entry.display_name : null) ?? entry.id,
-    }));
-}
-
 /**
  * Whether the roster this surface is showing stopped short of the workspace.
  *
